Refactored dh utils to not have dependencies on the extension (#79-1)
