--- src/util/ErrorTypes.ts
@@ -1,11 +1,0 @@
-export class InvalidConsoleTypeError extends Error {
-  constructor(type: string) {
-    super(`Invalid console type: '${type}'`);
-  }
-}
-
-export class NoConsoleTypesError extends Error {
-  constructor() {
-    super('No console types available');
-  }
-}
