+++ src/util/index.ts
@@ -1,16 +1,13 @@
 export * from './assertUtil';
 export * from './dataUtils';
-export * from './downloadUtils';
-export * from './errorUtils';
-export * from './ErrorTypes';
 export * from './isDisposable';
 export * from './Logger';
 export * from './OutputChannelWithHistory';
 export * from './panelUtils';
-export * from './polyfillUtils';
 export * from './selectionUtils';
 export * from './serverUtils';
 export * from './testUtils';
+export * from './tmpUtils';
 export * from './treeViewUtils';
 export * from './Toaster';
 export * from './uiUtils';
